# docs/swagger.yaml
basePath: /
definitions:
...
  model.Course:
    properties:
      codeCourse:
        type: string
      id:
        type: integer
      lessons:
        type: integer
      name:
        type: string
      subject:
        $ref: '#/definitions/model.Subject'
      subjectId:
        type: integer
      teacher:
        $ref: '#/definitions/model.Teacher'
      teacherId:
        type: integer
    type: object
...
  model.Student:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
  model.Subject:
    properties:
      codeSubject:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.Teacher:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Api canvas project
  title: Canvas Project
  version: "2.0"
paths:
  /admin/course/create:
    post:
      consumes:
      - application/json
      description: Admin create course
      parameters:
      - description: Course
        format: model.Course
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Course'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with course
  /admin/course/delete:
    delete:
      consumes:
      - application/json
      description: Admin delete course
      parameters:
      - description: Course need delete
        format: model.Course
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Course'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with course
...
  /admin/course/list/{id}:
    get:
      consumes:
      - application/json
      description: Admin get course by id
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with course
  /admin/course/update:
    put:
      consumes:
      - application/json
      description: Admin update course
      parameters:
      - description: Course need update
        format: model.Course
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Course'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with course
...
  /admin/student/accounts/{id}:
    get:
      consumes:
      - application/json
      description: Admin get account student by id
      parameters:
      - description: Student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account student
  /admin/student/create:
    post:
      consumes:
      - application/json
      description: Admin create account student
      parameters:
      - description: Account student
        format: model.Student
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account student
  /admin/student/delete:
    delete:
      consumes:
      - application/json
      description: Admin delete account student
      parameters:
      - description: Account stundent need delete
        format: model.Student
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account student
  /admin/student/update:
    put:
      consumes:
      - application/json
      description: Admin update account student
      parameters:
      - description: Account student need update
        format: model.Student
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account student
  /admin/subject/create:
    post:
      consumes:
      - application/json
      description: Admin create subject
      parameters:
      - description: Subject
        format: model.Subject
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Subject'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with subject
  /admin/subject/delete:
    delete:
      consumes:
      - application/json
      description: Admin delete subject
      parameters:
      - description: Subject need delete
        format: mode.Subject
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Subject'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with subject
...
  /admin/subject/list/{id}:
    get:
      consumes:
      - application/json
      description: Admin get subject by id
      parameters:
      - description: Subject id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with subject
  /admin/subject/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Subject need update
        format: model.Subject
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Subject'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with subject
...
  /admin/teacher/accounts/{id}:
    get:
      consumes:
      - application/json
      description: Admin get account teacher by id
      parameters:
      - description: Teacher Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account teacher
  /admin/teacher/create:
    post:
      consumes:
      - application/json
      description: Admin create account teacher
      parameters:
      - description: Account teacher
        format: model.Teacher
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Teacher'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account teacher
  /admin/teacher/delete:
    delete:
      consumes:
      - application/json
      description: Admin delete account teacher
      parameters:
      - description: Account teacher need delete
        format: model.Teacher
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Teacher'
      responses: {}
      tags:
      - Admin with account teacher
  /admin/teacher/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Account teacher need update
        format: model.Teacher
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Teacher'
      produces:
      - application/json
      responses: {}
      tags:
      - Admin with account teacher
...
swagger: "2.0"